changed the way ids work
Habits and Events there ids are now for local identification, the
unique Id Generator can make sure the ids are unique. for server
identification it uses the unique user id that has to be gotten from
the server plus the unique local id. this way the id used to push to
the sever is unique across all the devices because the server gives us
a unique user id on the first boot up